Create main_activity.java
In response to the layout we produced java files known as Main Activity, in this java file we defined the parameters included in the layout, such as text view and more. There in the file we made a function (method) to define what happens when we click those text views defined in content_main layout that further included in activity_main layout. As soon as we define the click listener, we defined where it goes meaning which activity class (another java file) is executed after this click.