Create map_destination.java
In  map_destination here we included map fragment and some more text views to increase understand-ability. And further we included some buttons and these buttons either reset the map or take you to another activity (more defined on map_destination class)	In response to the layout we produced a class map_destination where we as usual defined the buttons and map fragment too.
Now what these buttons do, the reset_button clears all the markers on the map and next_button takes you to the “Computation class” as soon as you selected your source and destination position.